requirements.txt
numpy
scipy
sounddevice
soundfile
matplotlib
rich
requests
python-dotenv
psycopg2-binary
